refactor middleware out of server.js
